examples/chat_app/requirements.txt
streamlit
cassandra-driver>=3.25.0
python-dotenv
langchain
langchain-openai